+++ assets/scripts/plugin.js
@@ -5,15 +5,29 @@ window.CookieNotice = ( function( window, document ) {
 		var Cookies = require('js-cookie');
 
 		if ( 'undefined' == typeof Cookies.get('cookie_notice') ) {
-			var el = document.createElement('div');
 
-			el.setAttribute( 'id', 'geniem-cookie-notice');
+			app.cookieNoticeElem = document.createElement('div');
+
+			app.cookieNoticeElem.setAttribute( 'id', 'geniem-cookie-notice' );
+			app.cookieNoticeElem.setAttribute( 'class', 'geniem-cookie-notice' );
+
+			var el = document.createElement( 'div' );
+
+			el.setAttribute( 'class', 'geniem-cookie-notice__inner');
 
 			var p = document.createElement('p');
+			p.setAttribute( 'class', 'geniem-cookie-notice__text-and-button' );
 
-			p.textContent = cookie_notice.cookie_notice_text;
+			var span = document.createElement( 'span' );
+
+			span.setAttribute( 'class', 'geniem-cookie-notice__text' );
+
+			span.textContent = cookie_notice.cookie_notice_text;
+
+			p.appendChild( span );
 
 			var button = document.createElement('button');
+			button.setAttribute( 'class', 'geniem-cookie-notice__button' );
 
 			button.textContent = cookie_notice.ok_text;
 
@@ -23,8 +37,11 @@ window.CookieNotice = ( function( window, document ) {
 
 			if ( cookie_notice.link_url ) {
 				var paragraph = document.createElement('p');
+				paragraph.setAttribute( 'class', 'geniem-cookie-notice__details' );
+
 				var link = document.createElement('a');
 
+				link.setAttribute( 'class', 'geniem-cookie-notice__link' );
 				link.setAttribute( 'href', cookie_notice.link_url );
 
 				link.textContent = cookie_notice.link_text;
@@ -34,18 +51,36 @@ window.CookieNotice = ( function( window, document ) {
 				el.appendChild( paragraph );
 			}
 
-			document.body.appendChild( el );
+			app.cookieNoticeElem.appendChild( el );
 
-			app.bodyMargin = el.offsetHeight + 'px';
+			document.body.appendChild( app.cookieNoticeElem );
 
-			document.body.style.marginBottom = app.bodyMargin;
+			// Get the original body margin, so that it may be resized later on.
+			app.originalBodyMarginBottom = window.getComputedStyle(document.body).getPropertyValue('margin-bottom');
+
+			app.setBodyMargin();
 
 			var cookieExpires = parseInt( cookie_notice.expires );
 
+			window.addEventListener('resize', app.setBodyMargin);
+
 			button.addEventListener( 'click', function( e ) {
 				Cookies.set( 'cookie_notice', true, { expires: cookieExpires } );
-				el.parentNode.removeChild( el );
+				app.cookieNoticeElem.parentNode.removeChild( app.cookieNoticeElem );
+
+				// Restore body bottom margin.
+				document.body.style.marginBottom = app.originalBodyMarginBottom;
+
+				// Remove event resize listener.
+				window.removeEventListener('resize', app.setBodyMargin);
 			});
+		}
+	};
+
+	// Set body margin to match the cookie notice element height.
+	app.setBodyMargin = function() {
+		if (app.cookieNoticeElem) {
+			document.body.style.marginBottom = app.cookieNoticeElem.offsetHeight + 'px';
 		}
 	};