+++ plugin.php
@@ -5,7 +5,7 @@
  * Author: Miika Arponen / Geniem Oy
  * Author URI: http://www.geniem.com
  * Text Domain: geniem-cookie-notice
- * Version: 0.0.6
+ * Version: 1.0.0
  */
 
 namespace Geniem;