+++ CHANGELOG.md
@@ -3,8 +3,15 @@ Plugin's changelog
 
 Based on a project [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
+## [1.0.0] - 2018-10-15
+
+### Modified
+- Added inner wrapping element to markup. Also added some classnames to elements.
+- Change body margin-bottom on resize.
+- Restore body margin-bottom when cookie notize element closed.
 
 ## [0.0.6] - 2018-03-09
+
 ### Modified
 - Fixed bug: changed cookie notice expires to int before passing it to Cookie.set.